+++ src/+aod/+builtin/+sources/+primate/Primate.m
@@ -1,21 +1,25 @@
 classdef Primate < aod.core.sources.Subject
-    % PRIMATE
-    %
-    % Description:
-    %   Subject class tailored for UR primates
-    %
-    % Parent:
-    %   aod.core.sources.Subject
-    %
-    % Constructor:
-    %   obj = Primate(name)
-    %
-    % Parameters:
-    %   DateOfBirth
-    %
-    % Dependent properties:
-    %   ID                      double, ID extracted from name
-    % ---------------------------------------------------------------------
+% Non-human primate
+%
+% Description:
+%   Subject class tailored for NHPs
+%
+% Parent:
+%   aod.core.sources.Subject
+%
+% Constructor:
+%   obj = aod.core.sources.Primate(name)
+%
+% Parameters:
+%   DateOfBirth
+%
+% Dependent properties:
+%   ID                      double, ID extracted from name
+%
+%! Move to sara-aodata-package
+
+% By Sara Patterson, 2023 (AOData)
+% -------------------------------------------------------------------------
     
     properties (Dependent)
         ID
@@ -24,10 +28,6 @@ classdef Primate < aod.core.sources.Subject
     methods
         function obj = Primate(name, varargin)
             obj = obj@aod.core.sources.Subject(name, varargin{:});
-
-            if ~obj.hasParam('DateOfBirth')
-                obj.setParam('Age', round(years(datetime('now') - obj.getParam('DateOfBirth')),1));
-            end
         end
 
         function value = get.ID(obj)
@@ -39,7 +39,8 @@ classdef Primate < aod.core.sources.Subject
         function value = specifyParameters(obj)
             value = specifyParameters@aod.core.sources.Subject(obj);
 
-            value.add('DateOfBirth', [], @isdatetime,...
+            value.remove('Age');
+            value.add('DateOfBirth', datetime.empty(), @isdatetime,...
                 'Date of birth of the subject');
         end
     end