+++ CHANGELOG.md
@@ -4,4 +4,4 @@ Documentation of changes prior to alpha release, beginning 30Apr2023
 
 #### 30-Apr-2023
 - Deprecated `aod.builtin.stimuli.VisualStimulus`
-  
+- Implemented expected parameter check in `aod.core.Entity` parameter methods